+++ CHANGELOG.md
@@ -9,3 +9,21 @@ This is the initial release of 1337 Translator.
 ### Goals for Next Release
 
  + Correct any bugs discovered and get user feedback from this release.
+
+## 0.2
+
+**Currently in the trunk.**
+
+### Bug Fixes
+
+ + Fixed only one output option working at a time
+
+### Tweaks
+
+ + ```--full-stop``` will no longer display a result output if there were no full stops to process.
+
+### New Features
+
+ + Text input can now be entered at the command line.
+ + The preprocessing display now has colors to aid in viewing the changes.
+ + There is now a version option to view the current version of this program.